refactor: tidy c_expr
