+++ crates/uroborosql-fmt/src/new_visitor/pg_expr/c_expr.rs
@@ -1,11 +1,12 @@
+mod aexpr_const;
+mod columnref;
 mod func_expr;
 
 use postgresql_cst_parser::{syntax_kind::SyntaxKind, tree_sitter::TreeCursor};
 
 use crate::{
-    cst::{AsteriskExpr, Comment, Expr, ParenExpr, PrimaryExpr, PrimaryExprKind},
+    cst::{Comment, Expr, ParenExpr},
     error::UroboroSQLFmtError,
-    util::convert_identifier_case,
 };
 
 use super::{pg_ensure_kind, pg_error_annotation_from_cursor, Visitor};
@@ -155,146 +156,6 @@ impl Visitor {
 
         cursor.goto_parent();
         pg_ensure_kind(cursor, SyntaxKind::c_expr, src)?;
-
-        Ok(expr)
-    }
-
-    // AexprConst
-    fn visit_aexpr_const(
-        &mut self,
-        cursor: &mut TreeCursor,
-        src: &str,
-    ) -> Result<Expr, UroboroSQLFmtError> {
-        // AexprConst
-        // - Iconst
-        //   - ICONST
-        // - FCONST
-        // - Sconst
-        //   - SCONST
-        // - BCONST
-        // - XCONST
-        // - func_name Sconst
-        // - func_name '(' func_arg_list opt_sort_clause ')' Sconst
-        // - ConstTypename Sconst
-        // - ConstInterval Sconst opt_interval
-        // - ConstInterval '(' Iconst ')' Sconst
-        // - TRUE_P
-        // - FALSE_P
-        // - NULL_P
-
-        cursor.goto_first_child();
-        let expr = match cursor.node().kind() {
-            SyntaxKind::Iconst
-            | SyntaxKind::FCONST
-            | SyntaxKind::Sconst
-            | SyntaxKind::BCONST
-            | SyntaxKind::XCONST => {
-                PrimaryExpr::with_pg_node(cursor.node(), PrimaryExprKind::Expr)?.into()
-            }
-            SyntaxKind::func_name => {
-                // func_name Sconst
-                // func_name '(' func_arg_list opt_sort_clause ')' Sconst
-                return Err(UroboroSQLFmtError::Unimplemented(format!(
-                    "visit_aexpr_const(): func_name is not implemented\n{}",
-                    pg_error_annotation_from_cursor(cursor, src)
-                )));
-            }
-            SyntaxKind::ConstTypename => {
-                // ConstTypename Sconst
-                return Err(UroboroSQLFmtError::Unimplemented(format!(
-                    "visit_aexpr_const(): ConstTypename is not implemented\n{}",
-                    pg_error_annotation_from_cursor(cursor, src)
-                )));
-            }
-            SyntaxKind::ConstInterval => {
-                // ConstInterval Sconst opt_interval
-                // ConstInterval '(' Iconst ')' Sconst
-                return Err(UroboroSQLFmtError::Unimplemented(format!(
-                    "visit_aexpr_const(): ConstInterval is not implemented\n{}",
-                    pg_error_annotation_from_cursor(cursor, src)
-                )));
-            }
-            SyntaxKind::TRUE_P | SyntaxKind::FALSE_P | SyntaxKind::NULL_P => {
-                PrimaryExpr::with_pg_node(cursor.node(), PrimaryExprKind::Keyword)?.into()
-            }
-            _ => {
-                return Err(UroboroSQLFmtError::UnexpectedSyntax(format!(
-                    "visit_aexpr_const(): unexpected node kind\n{}",
-                    pg_error_annotation_from_cursor(cursor, src)
-                )));
-            }
-        };
-
-        cursor.goto_parent();
-        pg_ensure_kind(cursor, SyntaxKind::AexprConst, src)?;
-
-        Ok(expr)
-    }
-
-    fn visit_columnref(
-        &mut self,
-        cursor: &mut TreeCursor,
-        src: &str,
-    ) -> Result<Expr, UroboroSQLFmtError> {
-        // columnref
-        // - ColId
-        // - ColId indirection
-        //   - e.g.: `a.field`, `a.field[1]`
-
-        // cursor -> ColId (必ず存在する)
-        cursor.goto_first_child();
-
-        pg_ensure_kind(cursor, SyntaxKind::ColId, src)?;
-        let mut columnref_text = cursor.node().text().to_string();
-
-        if cursor.goto_next_sibling() {
-            // cursor -> indirection
-            pg_ensure_kind(cursor, SyntaxKind::indirection, src)?;
-
-            // indirection
-            // - indirection_el
-            //    - `.` attr_name
-            //    - `.` `*`
-            //    - `[` a_expr `]`
-            //    - `[` opt_slice_bound `:` opt_slice_bound `]`
-            //
-            // indirection はフラット化されている: https://github.com/future-architect/postgresql-cst-parser/pull/7
-
-            let indirection_text = cursor.node().text();
-
-            // 配列アクセスは unimplemented
-            if indirection_text.contains('[') {
-                return Err(UroboroSQLFmtError::Unimplemented(format!(
-                    "visit_columnref(): array access is not implemented\n{}",
-                    pg_error_annotation_from_cursor(cursor, src)
-                )));
-            }
-
-            // indirection にあたるテキストから空白文字を除去し、そのまま追加している
-            let whitespace_removed = indirection_text
-                .chars()
-                .filter(|c| !c.is_whitespace())
-                .collect::<String>();
-            columnref_text.push_str(&whitespace_removed);
-        }
-
-        // アスタリスクが含まれる場合はAsteriskExprに変換する
-        let expr = if columnref_text.contains('*') {
-            AsteriskExpr::new(
-                convert_identifier_case(&columnref_text),
-                cursor.node().range().into(),
-            )
-            .into()
-        } else {
-            PrimaryExpr::new(
-                convert_identifier_case(&columnref_text),
-                cursor.node().range().into(),
-            )
-            .into()
-        };
-
-        cursor.goto_parent();
-        pg_ensure_kind(cursor, SyntaxKind::columnref, src)?;
 
         Ok(expr)
     }
